// src/app/core/store/feature.enum.ts
export enum StoreFeatureKey {
  MESSAGE = 'message',
  MESSAGE_MAIL_FOLDER = 'message-mail-folder',
  TASK = 'task',
  TASK_PREVIEW = 'taskPreview',
  CONVERSATION = 'conversation',
  TASK_FOLDER = 'taskFolder',
  CALENDAR_EVENT = 'calendarEvent',
  CALENDAR_DASHBOARD = 'calendarDashboard',
  TENANTS_OWNERS_PT = 'tenants-owners-pt',
  TENANTS_OWNERS_RM = 'tenants-owners-rm',
  SUPPLIER = 'supplier',
  OTHER_CONTACT = 'other-contact',
  TENANT_PROSPECT = 'tenant-prospect',
  OWNER_PROSPECT = 'owner-prospect',
  MAIL_FOLDER = 'mail-folder',
  APP_MESSAGE = 'app-messages',
  VOICE_MAIL = 'voice-mail',
  VOICE_MAIL_DETAIL = 'voice-detail',
  TASK_DETAIL = 'taskDetail',
  CONVERSATION_ACTION_DETAIL = 'conversation-action-detail',
  FACEBOOK = 'facebook',
  FACEBOOK_DETAIL = 'facebook-detail',
  SMS_MESSAGE = 'sms-message',
  WHATSAPP = 'whatsapp',
  WHATSAPP_DETAIL = 'whatsapp-detail'
}